HTML Home

+++ db/spending_tracker.sql
@@ -20,7 +20,7 @@ CREATE TABLE merchants(
 
 CREATE TABLE transactions(
   id SERIAL PRIMARY KEY,
-  amount INT,
+  amount DECIMAL(18,2),
   merchant_id INT REFERENCES merchants(id) ON DELETE CASCADE,
   tag_id INT REFERENCES tags(id) ON DELETE CASCADE
 );
